make reusable gallery
