+++ styles/gallery.css
@@ -1,0 +1,217 @@
+.gallery {
+    position: fixed;
+    top: 0;
+    left: 0;
+    right: 0;
+    bottom: 0;
+    background: rgba(0, 0, 0, 0.9);
+    z-index: 1000;
+    display: none;
+    overflow-y: auto;
+    padding: 60px 20px 20px;
+  }
+  
+  .gallery.active {
+    display: block;
+  }
+  
+  .gallery-close, .overlay-close {
+    position: fixed;
+    top: 20px;
+    right: 20px;
+    background: none;
+    border: none;
+    color: white;
+    font-size: var(--body-font-size-l);
+    cursor: pointer;
+    z-index: 1001;
+  }
+  
+  .gallery-content {
+    display: grid;
+    max-width: 1200px;
+    margin: 0 auto;
+    grid-auto-flow: row dense;
+    grid-gap: var(--gallery-gap);
+    grid-template-columns: repeat(3, 1fr);
+  }
+  
+  .gallery-item {
+    position: relative;
+    overflow: hidden;
+    cursor: pointer;
+    width: 100%;
+  }
+  
+  .gallery-item img {
+    width: 100%;
+    height: 100%;
+    object-fit: cover;
+  }
+  
+  .gallery-item:nth-child(9n+1),
+  .gallery-item:nth-child(9n+2) {
+    grid-column: 1;
+  }
+  
+  .gallery-item:nth-child(9n+3) {
+    grid-column: 2 / span 2;
+    grid-row: span 2;
+  }
+  
+  .gallery-item:nth-child(9n+4),
+  .gallery-item:nth-child(9n+5),
+  .gallery-item:nth-child(9n+6) {
+    grid-column: auto;
+  }
+  
+  .gallery-item:nth-child(9n+7) {
+    grid-column: 1 / span 2;
+    grid-row: span 2;
+  }
+  
+  .gallery-item:nth-child(9n+8),
+  .gallery-item:nth-child(9n+9) {
+    grid-column: 3;
+  }
+  
+  .gallery-item.large {
+    grid-column: span 2;
+    grid-row: span 2;
+  }
+  
+  .gallery-item.large img {
+    height: 354px;
+  }
+  
+  .gallery-item picture {
+    width: 100%;
+    display: block;
+  }
+  
+  .gallery-item img {
+    width: 100%;
+    height: auto;
+    object-fit: cover;
+    border-radius: 10px;
+  }
+  
+  .image-overlay {
+    position: fixed;
+    inset: 0;
+    background: rgba(0, 0, 0, 0.9);
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    z-index: 2000;
+  }
+  
+  .image-overlay-content {
+    position: relative;
+    max-width: 90%;
+    max-height: 90%;
+    width: 100%;
+    aspect-ratio: 16 / 9;
+  }
+  
+  .image-overlay img {
+    width: 100%;
+    height: 100%;
+    object-fit: contain;
+  }
+  
+  .image-overlay .btns {
+    position: absolute;
+    top: 0;
+    left: 0;
+    width: 100%;
+    height: 100%;
+  }
+  
+  .image-overlay .btns button {
+    position: absolute;
+    top: 0;
+    left: 0;
+    width: 25%;
+    height: 100%;
+    z-index: 3;
+    cursor: pointer;
+    background: transparent;
+    outline: none;
+    border: 0;
+    -webkit-tap-highlight-color: transparent;
+  }
+  
+  .image-overlay .btns button.next {
+    left: auto;
+    right: 0;
+  }
+  
+  .image-overlay .btns button::after {
+    position: absolute;
+    display: block;
+    content: '\e079';
+    top: 50%;
+    left: 20%;
+    z-index: 3;
+    width: 30px;
+    height: 30px;
+    margin-top: -15px;
+    font-family: glyphicons;
+    font-size: var(--heading-font-size-xl);
+    color: #fff;
+    line-height: 1;
+    text-shadow: 0 1px 2px rgba(0 0 0 / 60%);
+    opacity: .5;
+    transition: opacity 200ms;
+  }
+  
+  .image-overlay .btns button:focus::after {
+    opacity: .5;
+  }
+  
+  .image-overlay .btns button:hover::after,
+  .image-overlay .btns button:focus:hover::after {
+    opacity: 1;
+  }
+  
+  .image-overlay .btns button.next::after {
+    transform: rotate(180deg);
+    left: auto;
+    right: 80px;
+  }
+  
+  @media (width <= 1200px) {
+    .gallery-content {
+      grid-template-columns: repeat(2, 1fr);
+    }
+  
+    .gallery-item.large {
+      grid-column: span 2;
+      grid-row: span 2;
+    }
+  
+    .gallery-item.large img {
+      grid-column: span 2;
+      grid-row: span 2;
+    }
+  
+    .gallery-item:nth-child(9n+1),
+    .gallery-item:nth-child(9n+2),
+    .gallery-item:nth-child(9n+4),
+    .gallery-item:nth-child(9n+5) {
+      grid-column: auto;
+    }
+  
+    .gallery-item:nth-child(9n+3),
+    .gallery-item:nth-child(9n+7) {
+      grid-column: 1 / span 2;
+      grid-row: span 2;
+    }
+  
+    .gallery-item:nth-child(9n+6),
+    .gallery-item:nth-child(9n+8),
+    .gallery-item:nth-child(9n+9) {
+      grid-column: auto;
+    }
+  }
